Image_classifier_ResNet-main/requirements.txt
pandas
keras
tensorflow
fastapi
uvicorn
Pillow
matplotlib
python-multipart
